+++ app/gui/src/project-view/components/MarkdownEditor/MarkdownEditorImpl.vue
@@ -2,8 +2,7 @@
 import CodeMirrorRoot from '@/components/CodeMirrorRoot.vue'
 import { transformPastedText } from '@/components/DocumentationEditor/textPaste'
 import BlockTypeDropdown from '@/components/MarkdownEditor/BlockTypeDropdown.vue'
-import { ensoMarkdown } from '@/components/MarkdownEditor/markdown'
-import { useMarkdownFormatting } from '@/components/MarkdownEditor/markdown/formatting'
+import { ensoMarkdown, useMarkdownFormatting } from '@/components/MarkdownEditor/codemirror'
 import ToggleIcon from '@/components/ToggleIcon.vue'
 import VueHostRender, { VueHostInstance } from '@/components/VueHostRender.vue'
 import { useCodeMirror } from '@/util/codemirror'
