+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/decoration/table.ts
