+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/decoration/treeStateDecorator.ts
