+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/decoration/treeViewDecorator.ts
