+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/index.ts
@@ -1,7 +1,8 @@
-import { markdownDecorators } from '@/components/MarkdownEditor/markdown/decoration'
-import { markdownFormatting } from '@/components/MarkdownEditor/markdown/formatting'
+import { markdownDecorators } from '@/components/MarkdownEditor/codemirror/decoration'
+import { markdownFormatting } from '@/components/MarkdownEditor/codemirror/formatting'
 import { ensoMarkdownSyntax } from '@/components/MarkdownEditor/markdown/syntax'
 import { type Extension } from '@codemirror/state'
+export { useMarkdownFormatting } from '@/components/MarkdownEditor/codemirror/formatting'
 
 /**
  * CodeMirror Extension for the Enso Markdown dialect.
